feat: add memory limiter processor pipeline validation
- move otel/container to otel/index to represent central collector resource abstraction
- fix OTLP receiver grpc config
- convert config to yaml when creating container, to enable referencing config
